pyproject.toml
[project]
name = "vidaplus"
version = "0.1.0"
description = ""
authors = [
    {name = "Rauane Lima"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "sqlalchemy[asyncio] (>=2.0.40,<3.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pwdlib[argon2] (>=0.2.1,<0.3.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "fpdf (>=1.7.2,<2.0.0)",
    "uvicorn[standard] (>=0.34.2,<0.35.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-cov = "^6.0.0"
taskipy = "^1.14.1"
ruff = "^0.11.2"
pytest-asyncio = "^0.26.0"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'fastapi dev vidaplus/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=vidaplus -vv'
post_test = 'coverage html'
